requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.0.0
smbus2==0.5.2
waveshare-epd==1.1.0
